Add missing dependencies in Linux instructions
In the "installing from source" instructions for Linux, a few
development package dependencies are missing.  They are needed to
install `devtools`, namely: libcurl, and ssl.  There might be
others.

I can only verify the package names for Fedora, I have updated
the Ubuntu and Arch instructions based on the following:

- Ubuntu
  https://packages.ubuntu.com/bionic/libcurl-openssl1.0-dev
  https://packages.ubuntu.com/bionic/libssl-dev

- Arch (openssl is pulled in as a dependency)
  https://aur.archlinux.org/packages/libcurl-openssl-1.0/